python/requirements.txt
fastapi
uvicorn[standerd]
opencv-python
python-multipart
pillow